+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Input editor: option to seeks to current frame after rewind (#134)
 * Input editor: show state invalidation as gray area when threads changed
 * Add setting to not load movie metadata
+* Input editor: pasting while selecting a range repeats the paste frames in range
 
 ### Changed